move wizard to prepalert package, for flat design
