+++ initial_config.go
@@ -12,7 +12,7 @@ import (
 //go:embed initial_config.hcl
 var initialConfigTemplate string
 
-func GenerateInitialConfig(sqsQueueName string) ([]byte, error) {
+func GenerateInitialConfig(sqsQueueName string, orgName string) ([]byte, error) {
 	t, err := template.New("inital_config").Parse(initialConfigTemplate)
 	if err != nil {
 		return nil, err
@@ -21,6 +21,7 @@ func GenerateInitialConfig(sqsQueueName string) ([]byte, error) {
 	err = t.Execute(&buf, map[string]string{
 		"Version":      Version,
 		"SQSQueueName": sqsQueueName,
+		"OrgName":      orgName,
 	})
 	if err != nil {
 		return nil, err
