--- testdata/runtime_eval_context.json
@@ -1,82 +1,0 @@
-{
-  "event": {
-    "alert": {
-      "closed_at": 1473130092,
-      "created_at": 1473129912693,
-      "critical_threshold": 1.9588528112516932,
-      "duration": 5,
-      "id": "2bj...",
-      "is_open": true,
-      "metric_label": "MetricName",
-      "metric_value": 2.255356387321597,
-      "monitor_name": "MonitorName",
-      "monitor_operator": "\u003e",
-      "opened_at": 1473129912,
-      "status": "critical",
-      "trigger": "monitor",
-      "url": "https://mackerel.io/orgs/.../alerts/2bj...",
-      "warning_threshold": 1.4665636369580741
-    },
-    "event": "alert",
-    "host": {
-      "id": "22D4...",
-      "is_retired": false,
-      "memo": "",
-      "name": "app01",
-      "roles": [
-        {
-          "fullname": "Service: Role",
-          "role_name": "Role",
-          "role_url": "https://mackerel.io/orgs/.../services/...",
-          "service_name": "Service",
-          "service_url": "https://mackerel.io/orgs/.../services/..."
-        }
-      ],
-      "status": "working",
-      "type": "unknown",
-      "url": "https://mackerel.io/orgs/.../hosts/..."
-    },
-    "image_url": "https://mackerel.io/embed/public/.../....png",
-    "memo": "memo....",
-    "org_name": "Macker..."
-  },
-  "params": {},
-  "query_result": {
-    "hoge_result": {
-      "borderless_table": "  Name           Sign            Rating  \n------- ----------------------- ---------\n  A      The Good                   500  \n  B      The Very very Bad Man      288  \n  C      The Ugly                   120  \n  D      The Gopher                 800  \n",
-      "columns": [
-        "Name",
-        "Sign",
-        "Rating"
-      ],
-      "json_lines": "{\"Name\":\"A\",\"Rating\":\"500\",\"Sign\":\"The Good\"}\n{\"Name\":\"B\",\"Rating\":\"288\",\"Sign\":\"The Very very Bad Man\"}\n{\"Name\":\"C\",\"Rating\":\"120\",\"Sign\":\"The Ugly\"}\n{\"Name\":\"D\",\"Rating\":\"800\",\"Sign\":\"The Gopher\"}\n",
-      "markdown_table": "| Name |         Sign          | Rating |\n|------|-----------------------|--------|\n| A    | The Good              |    500 |\n| B    | The Very very Bad Man |    288 |\n| C    | The Ugly              |    120 |\n| D    | The Gopher            |    800 |\n",
-      "name": "hoge_result",
-      "query": "dummy",
-      "rows": [
-        [
-          "A",
-          "The Good",
-          "500"
-        ],
-        [
-          "B",
-          "The Very very Bad Man",
-          "288"
-        ],
-        [
-          "C",
-          "The Ugly",
-          "120"
-        ],
-        [
-          "D",
-          "The Gopher",
-          "800"
-        ]
-      ],
-      "table": "+------+-----------------------+--------+\n| NAME |         SIGN          | RATING |\n+------+-----------------------+--------+\n| A    | The Good              |    500 |\n| B    | The Very very Bad Man |    288 |\n| C    | The Ugly              |    120 |\n| D    | The Gopher            |    800 |\n+------+-----------------------+--------+\n",
-      "vertical_table": "********* 1. row *********\n  Name: A\n  Sign: The Good\n  Rating: 500\n********* 2. row *********\n  Name: B\n  Sign: The Very very Bad Man\n  Rating: 288\n********* 3. row *********\n  Name: C\n  Sign: The Ugly\n  Rating: 120\n********* 4. row *********\n  Name: D\n  Sign: The Gopher\n  Rating: 800\n"
-    }
-  }
-}
