--- testdata/table.txt
@@ -1,8 +1,0 @@
-+------+-----------------------+--------+
-| NAME |         SIGN          | RATING |
-+------+-----------------------+--------+
-| A    | The Good              |    500 |
-| B    | The Very very Bad Man |    288 |
-| C    | The Ugly              |    120 |
-| D    | The Gopher            |    800 |
-+------+-----------------------+--------+
